requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
pyodbc>=4.0.39
python-dotenv>=1.0.0
sqlalchemy>=2.0.0